% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dag_fct.R
\name{create_dag}
\alias{create_dag}
\title{Create a plain example of DAG for shiny project}
\usage{
create_dag(treatment, history, colrs)
}
\arguments{
\item{treatment}{Coefficient of treatment variable.}

\item{history}{Coefficient of history variable.}

\item{colrs}{List of colors.}
}
\value{
ggplot of DAG
}
\description{
Create a plain example of DAG for shiny project
}
