% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_itbl_fct.R
\name{create_itbl}
\alias{create_itbl}
\title{Create an interactive table}
\usage{
create_itbl(data)
}
\arguments{
\item{data}{Dataframe.}
}
\value{
Formatted \code{reactable} object.
}
\description{
Create an interactive table
}
