% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tbl_fct.R
\name{create_tbl}
\alias{create_tbl}
\title{Create a table using the \code{gt} package}
\usage{
create_tbl(data, title = "Title", subtitle = "Subtitle")
}
\arguments{
\item{data}{Dataframe}

\item{title}{String. Title for the table.}

\item{subtitle}{String. Subtitle for the table.}
}
\value{
Formatted \code{gt_tbl} object
}
\description{
Create a table using the \code{gt} package
}
