% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_dag.R, R/mod_dag_fct.R
\name{ggp_dag}
\alias{ggp_dag}
\title{Plot a DAG with ggplot and node names with subscripts}
\usage{
ggp_dag(dag, layout = "sugiyama")

ggp_dag(dag, layout = "sugiyama")
}
\arguments{
\item{dag}{A directed acyclic graph from the \code{dagitty} package}

\item{layout}{String. Name of the layout. Default is "sugiyama".}
}
\value{
ggplot object

ggplot object
}
\description{
Plot a DAG with ggplot and node names with subscripts.

Plot a DAG with ggplot and node names with subscripts.
}
\details{
The DAG will be plotted using \code{ggplot2} and the nodes' names can have
subscript as in mathematical notations. See the \code{source} below for
more details.

The DAG will be plotted using \code{ggplot2} and the nodes' names can have
subscript as in mathematical notations. See the \code{source} below for
more details.
}
