% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyTemp1-package.R
\docType{package}
\name{ShinyTemp1-package}
\alias{ShinyTemp1}
\alias{ShinyTemp1-package}
\title{ShinyTemp1: Shiny Template 1}
\description{
Shiny template with simulation, tables and plot. Primarily as a demonstration on how to interconnect modules and sub modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FrankLef/ShinyTemp1}
  \item Report bugs at \url{https://github.com/FrankLef/ShinyTemp1/issues}
}

}
\author{
\strong{Maintainer}: François Lefebvre Lefebvre \email{flefebvre01@hotmail.com}

}
\keyword{internal}
