% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyTemp1-package.R
\docType{package}
\name{ShinyTemp1}
\alias{ShinyTemp1}
\title{A template package to start shiny project}
\description{
The package sets up the mosst basic and relevant shiny package for
François Lefebvre.
}
\section{ShinyTemp1 functions}{

The functions ...
}

