% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim.R
\name{sim_dag}
\alias{sim_dag}
\title{Simulate a DAG}
\usage{
sim_dag(n = 1000, probH = 0.4, seed = 222)
}
\arguments{
\item{n}{Sample size}

\item{probH}{probability of latent variable \code{H}}

\item{seed}{Seed. Default is 222}
}
\value{
Data.frame with probabilities
}
\description{
Simulate a DAG
}
\examples{
sim_dag()
}
