% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_fct.R
\name{sim_binom}
\alias{sim_binom}
\title{Simulate data having spurious associations using \code{rbinom}}
\usage{
sim_binom(n = 1000, probH = 0.4, seed = NULL)
}
\arguments{
\item{n}{Sample size}

\item{probH}{probability of latent variable \code{H}}

\item{seed}{Seed. Default is 222}
}
\value{
Dataframe, summary of binomial distribution probabilities
}
\description{
Simulate data having spurious associations using \code{rbinom}
}
\examples{
sim_binom()
}
