% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fct.R
\name{sim_norm}
\alias{sim_norm}
\title{Simulate data having spurious associations using \code{rnorm}}
\usage{
sim_norm(n = 1000, meanH = 0.4, sdH = 0.1, seed = 222)
}
\arguments{
\item{n}{Sample size}

\item{meanH}{mean of latent variable \code{H}}

\item{sdH}{sd of latent variable \code{H}}

\item{seed}{Seed. Default is 222}
}
\value{
Data.frame with probabilities
}
\description{
Simulate data having spurious associations using \code{rnorm}
}
\examples{
sim_norm()
}
